fix: change tick time to 1s, add parking as route for better route recognition
